+++ inc/usb_desc.h
@@ -8,7 +8,7 @@
 
 #define MASS_SIZ_STRING_LANGID            4
 #define MASS_SIZ_STRING_VENDOR            38
-#define MASS_SIZ_STRING_PRODUCT           38
+#define MASS_SIZ_STRING_PRODUCT           48
 #define MASS_SIZ_STRING_SERIAL            26
 #define MASS_SIZ_STRING_INTERFACE         16